refactor(lexer): split `replay` module into multiple files (#26560)
Pure refactor. No code alterations, just moving things around.

- Split `replay`'s tests into a separate module. This reduces the size of the main file, and is a little faster to build, as the `tests.rs` module doesn't get parsed unless compiling tests.
- Re-order functions in `replay` following same convention as other `pipeline` modules - the single entry point `replay_is_keyword`  at the top, with callees lower down, grouped thematically.
